Add files via upload
Added the test case, a modified version of cavity tutorial case to include the custom boundary condition at 'walls'. 
The setup and codes are in src code.

The test case exists in the run directory of OpenFOAM.
And the setup and codes are located in $WM_PROJECT_USER_DIR (or simply the OpenFOAM user) directory.